requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-shortuuidfield==0.1.3
django-taggit==3.1.0
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==1.0.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
